In renderTextAtlas the x,y parms now tell where the text line is
renderTextAtlas used to render the text below the given x,y coordinates
this lead to inconsistent text rendering ,text placement was depended on
what string was being rendered.
Now it will treat the x,y coordinates as whre the text line is and all
letters in the string will be rendered relatively to that line.

renderToRect was updated to to give the correct x,y coordinates to
renderTextAtlas for the different text formats

Fixed a bug with calcTextRect not calculating the text size correctly.

added clipTextToRect to clip text to a given rect.